changes in StratP and btP forgotten, changes for vbt introduced here, ib part put in a separate file
