+++ py-trading-bot/core/btP.py
@@ -18,7 +18,8 @@ from core.stratP import StratPRD
 import core.indicators as ic
 from core.bt import BT, WQ
 
-from orders.models import retrieve_data, get_candidates, Excluded, check_hold_duration, Strategy
+from orders.ib import retrieve_data,check_hold_duration
+from orders.models import  get_candidates, Excluded, Strategy
 
 from trading_bot.settings import (RETARD_MAX_HOLD_DURATION, HIST_VOL_SLOW_MAX_CANDIDATES_NB,
                                  REALMADRID_MAX_CANDIDATES_NB, REALMADRID_DISTANCE)
@@ -31,7 +32,7 @@ from trading_bot.settings import (RETARD_MAX_HOLD_DURATION, HIST_VOL_SLOW_MAX_CA
 #We try to mutualize the calculation as much as possible to avoid re-calculating the same several times
 #The orders are handled somewhere else, only the candidates need to be determined here 
 class Presel(BT):
-    def __init__(self,**kwargs):
+    def __init__(self,use_IB,**kwargs):
  
         if kwargs.get("st",False):
             st=kwargs.get("st")
@@ -47,14 +48,14 @@ class Presel(BT):
             self.open_ind=st.open_ind
             self.volume_ind=st.volume_ind
             
-            self.symbols=self.close.columns.values
+            #self.symbols=self.close.columns.values
         else:
-            self.symbols=kwargs.get("symbols1")
+            self.actions=kwargs.get("actions1")
             self.period=kwargs.get("period1")
             
             self.high, self.low, self.close, self.open,self.volume,\
             self.high_ind, self.low_ind, self.close_ind, self.open_ind, self.volume_ind\
-            =retrieve_data(self.symbols,self.period)
+            =retrieve_data(self.actions,self.period,use_IB )
                 
         self.candidates=[[] for ii in range(len(self.close))]
         self.candidates_short=[[] for ii in range(len(self.close))]
@@ -69,7 +70,7 @@ class Presel(BT):
             self.ent11=st.entries
             self.ex11=st.exits
         else:
-            st=StratPRD(self.symbols,self.period,**kwargs)
+            st=StratPRD(self.actions,self.period,use_IB,**kwargs)
             st.strat_kama_stoch_matrend_bbands()
             
             self.ent11=st.entries
@@ -185,7 +186,7 @@ class Presel(BT):
 
 #As WQ but for production
 class WQPRD(WQ):
-    def __init__(self,**kwargs):    
+    def __init__(self,use_IB ,**kwargs):    
         if kwargs.get("st",False):
             st=kwargs.get("st")
             
@@ -200,14 +201,14 @@ class WQPRD(WQ):
             self.open_ind=st.open_ind
             self.volume_ind=st.volume_ind
             
-            self.symbols=self.close.columns.values
+            #self.symbols=self.close.columns.values
         else:
-            self.symbols=kwargs.get("symbols1")
+            self.actions=kwargs.get("actions1")
             self.period=kwargs.get("period1")
             
             self.high, self.low, self.close, self.open,self.volume,\
             self.high_ind, self.low_ind, self.close_ind, self.open_ind, self.volume_ind\
-            =retrieve_data(self.symbols,self.period)
+            =retrieve_data(self.actions,self.period,use_IB )
                 
         self.candidates=[[] for ii in range(len(self.close))]
         self.trend=ic.VBTBBANDSTREND.run(self.close_ind).trend
